+++ packages/blocks/core/CHANGELOG.md
@@ -94,6 +94,7 @@
 - `Details` block: support improved.
 - `File` block: support improved.
 - `Gallery` block: support improved.
+- `Group` block: support improved.
 
 
 ### Automated Tests
@@ -125,6 +126,7 @@
 - Updated E2E tests to check support of `Details` block.
 - Updated E2E tests to check support of `File` block.
 - Updated E2E tests to check support of `Gallery` block.
+- Updated E2E tests to check support of `Group` block.
 
 
 ## 1.3.0 (2025-04-27)